LayoutTests/fast/events/mouse-click-events-expected.txt

This tests what mouse events we send.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Left Mouse Button
PASS eventLog is "mousedown(button: 0 which: 1) mouseup(button: 0 which: 1) click(button: 0 which: 1) mousedown(button: 0 which: 1) mouseup(button: 0 which: 1) click(button: 0 which: 1) dblclick(button: 0 which: 1) "
Middle Mouse Button
PASS eventLog is "mousedown(button: 1 which: 2) mouseup(button: 1 which: 2) auxclick(button: 1 which: 2) mousedown(button: 1 which: 2) mouseup(button: 1 which: 2) auxclick(button: 1 which: 2) "
Right Mouse Button
PASS eventLog is "mousedown(button: 2 which: 3) mouseup(button: 2 which: 3) auxclick(button: 2 which: 3) mousedown(button: 2 which: 3) mouseup(button: 2 which: 3) auxclick(button: 2 which: 3) "
Back Mouse Button
PASS eventLog is "mousedown(button: 3 which: 4) mouseup(button: 3 which: 4) auxclick(button: 3 which: 4) mousedown(button: 3 which: 4) mouseup(button: 3 which: 4) auxclick(button: 3 which: 4) "
Forward Mouse Button
PASS eventLog is "mousedown(button: 4 which: 5) mouseup(button: 4 which: 5) auxclick(button: 4 which: 5) mousedown(button: 4 which: 5) mouseup(button: 4 which: 5) auxclick(button: 4 which: 5) "
PASS successfullyParsed is true

TEST COMPLETE
